requirements.txt
streamlit
pymupdf
pillow
ocrmypdf
googletrans
nest_asyncio
openai
deepl
surya-ocr